Add a mechanism to make certain data available through whole app.
The default idea is to store username and permissions of the logged in user
as well as a state of the application and make that data available through
the whole application.
For now only the mechanism is pressent, no data is actually stored. When
this will be completed, the change will be visible in every window.